Double pushing pawns and capturing them would put engine in a bad state
A bug causing the engine to enter a bad state when double pushing pawns and capturing them has been addressed.

The order of operations in the MakeMove function was adjusted to ensure correct handling of ply count.

Additionally, some unnecessary code related to game context was removed for cleaner implementation.

Updated version from 0.6.2 -> 0.6.3